New Queue API.
RTOS_Enqueue() can now block (timeout is a parameter) if the queue is
full.

Also a push operation called RTOS_PrependQueue() is now supported to
push a previously removed element back to the queue or to use a queue as
a LIFO (stack).

RTOS_PeekQueue() returns a status like all other functions, an output
parameter is now necessary to pass the element back to the caller.

Adjusted example to use the new parameter format.