Scheduler lock behavior.
I am seriously thinking about deprecating scheduler locking and perhaps
removing it altogether.
This change at least makes it behave and does not allow scheduler
locking in SMP and from inside an ISR.
Also bug fixes to RTOS_SuspendTask() and RTOS_KillTask() so they do not
return without getting out of a critical section. Also they cannot work
if the scheduler is locked and we are trying to suspend or kill the
current task. Added a check for that.