Adding an X86 target for demo purposes.
While JaeOS is aimed at small embedded systems, not a desktop compyter,
the most common type of hardware that people have for testing is an X86
based PC. So I am adding a target for that. It is probably more of a
demo target as any realistic system would need a lot of work to be
suitable for general use.