Another incarnation of sys_arch.c for interfacing to lwIP.
This is another incarnation of sys_arch.c.
This version will create a non-time share task if the priority passed to
sys_thread_new() is not the default one. The new thread will have the
priority specified.
This can be used e.g. to make the TCP/IP and network input threads
higher priority than the normal application threads.