Save and restore Neon context on interrupts and task switch.
It seems that Xilinx tools by default assume that Neon SIMD is usable
even for non-floating point operations when optimization is turned on.
Also some precompiled libraries already use Neon.
Thus it is necessary to save and restore Neon context when seriicing
interrupts or as part of a task switch.